Decompose Blacklight::Utils
